requirements.txt
streamlit==1.18.1
pandas==1.5.1
#selenium==4.6.0
seleniumbase
webdriver-manager==3.8.4
openpyxl==3.0.10
